Merge in changes with 10a286
When Apple released Snow Leopard as a beta to developers, just before
abandoning ZFS, there were a number of improvements made to the ZFS
runtime. These bits are referred to as 10a286, presumably with reference
to some project build number at the time.

One developer requested that a source copy of the CDDL covered code
be released, and Apple agreed, providing just the CDDL licensed code
(but not the corresponding Apple-licensed libraries).

As a result, this is a reference stream that may be useful for others to
introspect, or to perform comparisons against the pre_merge_72 tag,
in case this helps solve currently existing problems in the future.
It is unlikely to compile without direct investment in the missing
files, and it may be the case that the Apple stream is a natural
dead end. It is being provided for reference purposes only.