Fixes trash issue.
OSX includes the '.' and '..' as directories in the count, so
anything that tries to list a directory will see both of these
as well. This confuses the Finder when emptying trash because
it counts just the files that are there. So only report on
attributes that are avaialble for real files.

Fixed by Jason R. McNeil.

Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>