Allow the release script to install
As well as being invoked with --debug, you can now
also invoke the release script with --install, which
will execute a 'sudo installer' on the package.

Any password prompts are raised by sudo; if you have a
recent sudo ticket then this will not be required; also
if you have an appropriate entry in the 'sudoers' file,
you won't be prompted.