Use single set of ZFS IOC numbers
The original Apple implementation used a collection of
ifdef values for the ZFS IOC commands. These got out of
step in onnv_75 as a new item was inserted in the middle
of the list (which incidentally, breaks backward compatibility
with older kernels).

Resolving down to a single list for upstream and downstream
was a challenge, since it involved a partially missing
set of defines and ordering of includes in order to work.

As a result, if future files have compile time errors complaining
of missing ZFS_IOC constants, check that the "libzfs_ioctl" is
included before the "libzfs_impl"

Issue: 57