Check symbols are defined for kext after build
The kernel can be compiled, but be missing symbols
which aren't resolvable once installed into the
main kernel. We can use 'kextutil' to verify whether
the transitive closure of symbols in conjunction with
the current running kernel in order to fail fast if
we detect that the kernel won't be able to load.