Added panic-decode to support investigation
Panic logs are specific to the machine in question,
along with the version of the kernel used at any
point. To make it easier to submit reports on
MacZFS, this utility will attempt to understand
the kernel panic log from the context of the
current kernel, machine, hardware and ZFS
version.

This has been tested on Intel i386 and should work
on x86_64. It doesn't work on PPC at the moment because
the backtrace is represented differently, and the 'kextutil'
used is only available on 10.6 systems.